+++ src/torch_pointcloud/models/spvcnn.py
@@ -1186,7 +1186,7 @@ def _spvcnn_semantickitti_hparams(cr: float) -> dict:
     weights=WeightsDict(
         url="hf://torch-pointcloud/spvcnn-30gmacs.semantickitti.mit-han-lab/resolve/main/model.safetensors",
         dataset="semantickitti",
-        metrics={"mIoU": 60.70},
+        metrics={"mIoU": 60.70, "OA": 91.94},
         classes=SEMANTIC_KITTI_CLASSES,
         author="mit-han-lab",
         license="MIT",
@@ -1204,7 +1204,7 @@ def spvcnn_30gmacs_semantickitti_seg(**hparams: Any) -> SPVCNNSegmentation:
     weights=WeightsDict(
         url="hf://torch-pointcloud/spvcnn-47gmacs.semantickitti.mit-han-lab/resolve/main/model.safetensors",
         dataset="semantickitti",
-        metrics={"mIoU": 61.59},
+        metrics={"mIoU": 61.59, "OA": 92.22},
         classes=SEMANTIC_KITTI_CLASSES,
         author="mit-han-lab",
         license="MIT",
@@ -1222,7 +1222,7 @@ def spvcnn_47gmacs_semantickitti_seg(**hparams: Any) -> SPVCNNSegmentation:
     weights=WeightsDict(
         url="hf://torch-pointcloud/spvcnn-119gmacs.semantickitti.mit-han-lab/resolve/main/model.safetensors",
         dataset="semantickitti",
-        metrics={"mIoU": 63.74},
+        metrics={"mIoU": 63.74, "OA": 92.46},
         classes=SEMANTIC_KITTI_CLASSES,
         author="mit-han-lab",
         license="MIT",